+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Created new component `Bulma::ButtonsComponent`
 - Updated `Elements::ButtonComponentPreview` to pass content as parameter
 - Added support to specify value as parameter to `Bulma::ButtonComponent`
 - Created `Bulma::DeleteComponent` with preview class and helper